# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Spector Cognitive Memory & Search API
  description: "Ultra-fast cognitive memory, vector search, and agent orchestration\
    \ platform with biological memory tiers (Working, Episodic, Semantic, Procedural)."
  contact:
    name: Spectrayan Maintainers
    url: https://github.com/spectrayan/spector
    email: support@spectrayan.com
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: /
  description: Current Server Instance
security:
- ApiKeyAuth: []
- BearerAuth: []
...
paths:
...
  /api/v1/agent/{id}:
    get:
      tags:
      - agent-controller
      operationId: getAgent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgentSoul"
    put:
      tags:
      - agent-controller
      operationId: updateAgent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentSoul"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgentSoul"
    delete:
      tags:
      - agent-controller
      operationId: deleteAgent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/agents/{id}:
    get:
      tags:
      - agent-controller
      operationId: getAgent_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgentSoul"
    put:
      tags:
      - agent-controller
      operationId: updateAgent_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentSoul"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgentSoul"
    delete:
      tags:
      - agent-controller
      operationId: deleteAgent_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /api/v1/agent/approvals/{id}/reject:
    post:
      tags:
      - agent-approval-controller
      operationId: reject
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/agent/approvals/{id}/modify:
    post:
      tags:
      - agent-approval-controller
      operationId: modify
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/agent/approvals/{id}/cancel:
    post:
      tags:
      - agent-approval-controller
      operationId: cancel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/agent/approvals/{id}/approve:
    post:
      tags:
      - agent-approval-controller
      operationId: approve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/agent/approvals:
    get:
      tags:
      - agent-approval-controller
      operationId: listApprovals
      parameters:
      - name: pendingOnly
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentActionApproval"
  /api/v1/agent/approvals/{id}:
    get:
      tags:
      - agent-approval-controller
      operationId: getApproval
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgentActionApproval"
...
components:
  schemas:
...
    AgentSoul:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        systemPrompt:
          type: string
        purpose:
          type: string
        personality:
          type: string
        expertiseDomains:
          type: array
          items:
            type: string
        coreValues:
          type: array
          items:
            type: string
        ethicalGuardrails:
          type: array
          items:
            type: string
        emotionalBaseline:
          $ref: "#/components/schemas/EmotionalBaseline"
        communicationStyle:
          type: string
        model:
          type: string
        tools:
          type: array
          items:
            type: string
        expertiseEmbedding:
          type: array
          items:
            type: number
            format: float
        purposeEmbedding:
          type: array
          items:
            type: number
            format: float
        soulVersion:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        present:
          type: boolean
    EmotionalBaseline:
      type: object
      properties:
        defaultValence:
          type: string
          format: byte
        defaultArousal:
          type: string
          format: byte
...
    ApprovalDecisionRequest:
      type: object
      properties:
        decision:
          type: string
          enum:
          - APPROVE
          - REJECT
          - MODIFY
          - CANCEL
        modifiedArguments:
          type: object
          additionalProperties: {}
        reason:
          type: string
...
    AgentActionApproval:
      type: object
      properties:
        id:
          type: string
        toolName:
          type: string
        arguments:
          type: object
          additionalProperties: {}
        category:
          type: string
        sessionId:
          type: string
        agentId:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
          - MODIFIED
          - CANCELLED
          - TIMED_OUT
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
        modifiedArguments:
          type: object
          additionalProperties: {}
        reason:
          type: string
        pending:
          type: boolean
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Spector API Key Authentication
      name: X-API-Key
      in: header
    BearerAuth:
      type: http
      description: OAuth2 / OIDC JWT Bearer Token
      scheme: bearer
      bearerFormat: JWT